Extract LocationSelector component
